test: Rename testutil.ts to httpserver.ts
